+++ R/filter_discrete.R
@@ -235,14 +235,16 @@ plot_feedback_bar <- function(plot_data, n_missing) {
   if (NROW(feedback_data) == 0) {
     gg_object <- ggplot2::ggplot()
   } else {
-
     gg_object <-
       feedback_data %>%
+      dplyr::mutate(
+        tooltip = htmltools::htmlEscape(paste0(level, " (", format_number(n), ")"), TRUE)
+      ) |>
       ggplot2::ggplot(
         ggplot2::aes(
           x = "I", y = n, fill = level,
           tooltip = paste0(level, " (", format_number(n), ")"),
-          data_id = level
+          data_id = htmltools::htmlEscape(level, TRUE)
         )
       ) +
       ggplot2::geom_col(position = ggplot2::position_stack(reverse = TRUE)) +