+++ NEWS.md
@@ -1,5 +1,9 @@
-# shinyCohortBuilder (development version)
+# shinyCohortBuilder 0.3.0
 
+* Most of the modals have now `easyClosed` option turned on (#25).
+* Fix positioning of filter help modal buttons (#37).
+* Fix rendering feedback plots when special character are included in data (#58).
+* `demo_app` can now return `shiny::shinyApp` object (#56)
 * Added new filter of type `query` based on 'shinyQueryBuilder' widget.
 * Added customization for filtering panel button classes via `scb_button_type` option.
 * Added customization for filtering panel labels via `scb_labels` option.